Added project in a satructure
