+++ BACKEND/dependency.py
@@ -1,0 +1,11 @@
+
+from database import SessionLocal
+
+
+# Dependency to get the database session
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
